+++ CAT/dye/addlig.py
@@ -10,35 +10,34 @@ Index
 .. autosummary::
     add_ligands
     export_dyes
+    sa_scores
 
 API
 ---
 .. autofunction:: add_ligands
 .. autofunction:: export_dyes
+.. autofunction:: sa_scores
 
 """
 
 import os
-from itertools import chain
-from typing import Iterator, Iterable, Collection
-from os.path import join, exists
-import numpy as np
-import pickle
 import gzip
 import math
+import pickle
+from typing import Iterator, Iterable, Collection
+from os.path import join, exists
+from itertools import chain
 
+import numpy as np
 from scm.plams import read_molecules, Molecule
 from scm.plams.interfaces.molecule.rdkit import to_rdmol
+from rdkit.Chem import rdMolDescriptors, FindMolChiralCenters
 
 from CAT.logger import logger
-
-from rdkit.Chem import AllChem, rdMolDescriptors, FindMolChiralCenters
-
-import CAT
 from CAT.attachment.dye import label_lig, label_core, substitution
 from CAT.attachment.substitution_symmetry import del_equiv_structures
 
-__all__ = ['add_ligands', 'export_dyes', 'SA_scores']
+__all__ = ['add_ligands', 'export_dyes', 'sa_scores']
 
 
 def add_ligands(core_dir: str,
@@ -143,8 +142,7 @@ def export_dyes(mol_list: Iterable[Molecule],
             )
 
 
-
-# The functions '_compute_SAS' and 'SA_scores' as well as data set 'SA_score.pkl.gz' 
+# The functions '_compute_SAS' and 'SA_scores' as well as data set 'SA_score.pkl.gz'
 # are copied and adapted from:
 ###########################################################################
 #    Title: MolGAN: An implicit generative model for small molecular graphs
@@ -154,10 +152,7 @@ def export_dyes(mol_list: Iterable[Molecule],
 ###########################################################################
 
 
-SA_model = {i[j]: float(i[0])
-            for i in pickle.load(gzip.open('SA_score.pkl.gz')) for j in range(1, len(i))}
-
-def _compute_SAS(mol):
+def _compute_sas(mol, sa_model: dict):
     fp = rdMolDescriptors.GetMorganFingerprint(mol, 2)
     fps = fp.GetNonzeroElements()
     score1 = 0.
@@ -166,7 +161,7 @@ def _compute_SAS(mol):
     for bitId, v in fps.items():
         nf += v
         sfp = bitId
-        score1 += SA_model.get(sfp, -4) * v
+        score1 += sa_model.get(sfp, -4) * v
     score1 /= nf
 
     # features score
@@ -194,8 +189,7 @@ def _compute_SAS(mol):
     if nMacrocycles > 0:
         macrocyclePenalty = math.log10(2)
 
-    score2 = 0. - sizePenalty - stereoPenalty - \
-             spiroPenalty - bridgePenalty - macrocyclePenalty
+    score2 = 0. - sizePenalty - stereoPenalty - spiroPenalty - bridgePenalty - macrocyclePenalty
 
     # correction for the fingerprint density
     # not in the original publication, added in version 1.1
@@ -221,10 +215,14 @@ def _compute_SAS(mol):
     return sascore
 
 
-def SA_scores(mols, norm=False):
-    mols = [to_rdmol(mol) for mol in mols]
-    scores = [_compute_SAS(mol) if mol is not None else None for mol in mols]
-    scores = np.array(list(map(lambda x: 10 if x is None else x, scores)))
-    scores = np.clip(remap(scores, 5, 1.5), 0.0, 1.0) if norm else scores
+def _load_sa_model(filename: str):
+    sa_score = pickle.load(gzip.open(filename))
+    return {i[j]: float(i[0]) for i in sa_score for j in range(1, len(i))}
 
-    return scores
+
+def sa_scores(mols: Iterable[Molecule], filename: str = 'SA_score.pkl.gz') -> np.ndarray:
+    """Calculate the synthetic accessibility score for all molecules in **mols**."""
+    sa_model = _load_sa_model(filename)
+    mols = (to_rdmol(mol) for mol in mols)
+    _scores = (_compute_sas(mol, sa_model) if mol is not None else None for mol in mols)
+    return np.array(list(map(lambda x: 10 if x is None else x, _scores)))